Bug fixes, consistency and quality of life
Added ids to the ui for the flags in 'index.html'.
Fixed a bug were no mines were placed in 'ts/main.ts',  made annotations more consistent, removed unnecessary code and improved the first click on the field, the tiles surrounding the click are now guaranteed to have no mines.