Bug fixes, comment improvement and more functionality
Added a line for css for debugging.
Changed the title in the html file.
Fixed some comments in "ts/main.ts", fixed a bug for setting the numbers in game and added functionality when clicking on a number when enough flags are surrounding it adjacent tiles are revealed.
And recompiled.