Major changes
Improved "compile.py" to remove the first two lines from "js/main.js".
Split up the style sheets into multiple files and improved styling.
Removed unnecessary code from "index.html" moved style to css files and added a part for the leader board.
Added two classes for the leader board to "ts/main.ts" and made it so that if a game is lost, all tiles are revealed. That functionality also needs a redo.
And recompiled.